Fix notification bugs, cleanup DownloadService + DownloadReceiver
This change started out just fixing a few regressions related to
notifications:

* Browser downloads weren't picking up a title from the determined
  filename.  This was due to my change to default the title field to
  "" instead of null.

* Notification click/hide events weren't being handled properly.  This
  was due to previous change to the URI structure of DownloadProvider.
  DownloadReceiver needed to be changed to perform queries through
  /all_downloads URIs, like all other parts of the download manager
  code.  I did some general refactoring of the DownloadReceiver code
  while I was there.

* The code in DownloadNotification wasn't picking up some updates to
  downloads properly.  This was due to my change to make
  DownloadNotification use the DownloadInfo objects rather than
  querying the database directly, so that it could make use of info
  provided by the DownloadThread that didn't go into the DB.  Fixing
  this didn't turn out to be all that complicated, but along the way
  to figuring this out I made some substantial refactoring in
  DownloadService which made it much cleaner anyway and eliminated a
  lot of duplication.  That's something that had to happen eventually,
  so I'm leaving it all in.

+++ src/com/android/providers/downloads/RealSystemFacade.java
@@ -81,13 +81,18 @@ class RealSystemFacade implements SystemFacade {
     }
 
     @Override
-    public void postNotification(int id, Notification notification) {
-        mNotificationManager.notify(id, notification);
+    public void postNotification(long id, Notification notification) {
+        /**
+         * TODO: The system notification manager takes ints, not longs, as IDs, but the download
+         * manager uses IDs take straight from the database, which are longs.  This will have to be
+         * dealt with at some point.
+         */
+        mNotificationManager.notify((int) id, notification);
     }
 
     @Override
-    public void cancelNotification(int id) {
-        mNotificationManager.cancel(id);
+    public void cancelNotification(long id) {
+        mNotificationManager.cancel((int) id);
     }
 
     @Override
